+++ docs/history.md
@@ -29,6 +29,7 @@
   - Necropolis: #10023, #10998, #20023, #20998
   - Mausoleum: #10248, #20248
 - Fixed the bug where the "Recent Achievements" menu would sometimes be in the reverse order.
+- Fixed the bug where pills from champions would not have proper seeds. (Thanks Neonomi)
 
 ### _v0.9.2_ - October 23rd, 2023